Switched Flyway for Liquibase along with some new database changes.

+++ src/main/java/io/skymind/pathmind/constants/PathmindConstants.java
@@ -3,5 +3,5 @@ package io.skymind.pathmind.constants;
 public class PathmindConstants
 {
 	// TODO -> Remove and implement
-	public static final long TODO_PARAMETER = 1;
+	public static final int TODO_PARAMETER = 1;
 }
