+++ src/main/java/io/skymind/pathmind/db/ExperimentRepository.java
@@ -1,15 +1,14 @@
 package io.skymind.pathmind.db;
 
 import io.skymind.pathmind.data.Experiment;
-import io.skymind.pathmind.data.db.tables.Project;
 import org.jooq.DSLContext;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
 
 import static io.skymind.pathmind.data.db.tables.Experiment.EXPERIMENT;
+
 
 @Repository
 public class ExperimentRepository
@@ -17,7 +16,7 @@ public class ExperimentRepository
 	@Autowired
 	private DSLContext dslContext;
 
-	public List<Experiment> getExperimentsForProject(long projectId) {
+	public List<Experiment> getExperimentsForProject(int projectId) {
 		return dslContext
 				.selectFrom(EXPERIMENT)
 				.where(EXPERIMENT.PROJECT_ID.eq(projectId))
